# src/GovUK.Dfe.FlexForms.Api.Client/Generated/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Api",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/tenant-config/tenants/{tenantId}": {
      "get": {
        "tags": [
          "TenantConfig"
        ],
        "summary": "Returns merged configuration for the specified tenant. For platform callers (shared Web container)\nthat resolve the tenant from the request hostname or X-Tenant-ID.",
        "operationId": "TenantConfig_GetTenantConfigurationByTenantId",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "target",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "Web"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantConfigurationDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid target.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Missing Platform.TenantConfig.Read app role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Tenant inactive or no longer exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenant-config/resolve": {
      "get": {
        "tags": [
          "TenantConfig"
        ],
        "summary": "Resolves a tenant id from an HTTP hostname using TenantHostnames.",
        "operationId": "TenantConfig_ResolveTenantByHostname",
        "parameters": [
          {
            "name": "hostname",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant resolved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResolutionDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid hostname.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Missing Platform.TenantConfig.Read app role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Hostname not mapped to any tenant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExceptionResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "errorId": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "exceptionType": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "correlationId": {
            "type": "string",
            "nullable": true
          },
          "context": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {}
          }
        }
      },
...
      "TenantConfigurationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "tenantId": {
            "type": "string",
            "format": "guid"
          },
          "tenantName": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "loadedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "configuration": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            }
          }
        }
      },
      "TenantResolutionDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "tenantId": {
            "type": "string",
            "format": "guid"
          },
          "tenantName": {
            "type": "string"
          },
          "hostname": {
            "type": "string"
          }
        }
      },
...
    }
  }
}